requirements.txt
numpy
pandas
opencv-python
scikit-learn
matplotlib
seaborn
tensorflow-cpu==2.12.0
keras
tqdm
